pyproject.toml
[project]
name = "audio-tuto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-cognitiveservices-speech>=1.43.0",
    "openai[realtime]>=1.72.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "simpleaudio>=1.0.4",
]
